requirements.txt
streamlit
langchain
python-dotenv
requests
groq
openai
